requirements.txt
aiodns==1.1.1
aiohttp==3.0.1
aioredis==1.1.0
alembic==0.9.9
amqp==2.2.2
appnope==0.1.0
asgiref==2.3.2
asn1crypto==0.24.0
async-generator==1.9
async-timeout==2.0.1
attrdict==2.0.0
attrs==18.1.0
autobahn==18.9.2
Automat==0.6.0
Babel==2.6.0
backcall==0.1.0
bcolz==1.2.1
billiard==3.5.0.3
bleach==2.1.3
boto3==1.5.27
botocore==1.8.50
Bottleneck==1.2.1
cchardet==2.1.1
ccxt==1.17.132
celery==4.2.0
certifi==2018.1.18
cffi==1.11.5
cfscrape==1.9.5
channels==2.0.2
channels-redis==2.3.0
chardet==3.0.4
click==6.7
coinbase-commerce==1.0.1
confusable-homoglyphs==3.1.1
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.3
cssselect==1.0.3
cycler==0.10.0
cyordereddict==1.0.0
Cython==0.27.3
cytoolz==0.9.0.1
daphne==2.2.2
decorator==4.3.0
defusedxml==0.5.0
Django==1.11.20
django-allauth==0.36.0
django-appconf==1.0.2
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-compressor==2.2
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-encrypted-fields-python3==1.1.3
django-extensions==2.1.2
django-extra-views==0.11.0
django-haystack==2.8.1
django-localflavor==2.1
django-oscar-paypal==1.0.0
django-phonenumber-field==2.0.0
django-redis==4.9.0
django-registration==2.4.1
django-tables2==1.21.2
django-treebeard==4.3
django-werkzeug-debugger-runserver==0.3.1
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
dnspython==1.15.0
docutils==0.14
empyrical==0.2.1
enigma-catalyst==0.5.15
entrypoints==0.2.3
eth-abi==1.1.1
eth-account==0.2.2
eth-hash==0.1.2
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
eventlet==0.24.1
factory-boy==2.11.1
Faker==0.8.17
flower==0.9.2
future==0.16.0
gevent==1.2.2
greenlet==0.4.13
h2==3.0.1
hexbytes==0.1.0
hiredis==0.2.0
hpack==3.0.0
html5lib==1.0.1
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.6
idna-ssl==1.1.0
incremental==17.5.0
intervaltree==2.1.0
ipdb==0.11
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
itchat==1.3.10
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
json-rpc==1.11.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterhub==0.9.0
kiwisolver==1.0.1
kombu==4.2.1
llvmlite==0.24.0
Logbook==0.12.5
lru-dict==1.1.6
lxml==4.2.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mock==2.0.0
monotonic==1.5
mpl-finance==0.10.0
msgpack==0.5.6
multidict==4.3.1
multipledispatch==0.4.9
mysqlclient==1.3.13
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook>=5.7.2
numba==0.39.0
numexpr==2.6.4
numpy==1.14.0
oauthenticator==0.7.3
oauthlib==2.1.0
pamela==0.3.0
pandas==0.19.2
pandas-datareader==0.6.0
pandocfilters==1.4.2
parsel==1.4.0
parsimonious==0.8.0
parso==0.2.0
patsy==0.5.0
pbr==4.1.0
pexpect==4.5.0
phonenumbers==8.9.10
phonenumberslite==8.9.10
pickleshare==0.7.4
Pillow==5.2.0
priority==1.3.0
prometheus-client==0.2.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
purl==1.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAthena==1.2.5
pycares==2.3.0
pycountry==18.5.26
pycparser==2.18
pycryptodome==3.6.6
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
pymongo==3.7.1
PyMySQL==0.8.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
pypng==0.0.18
PyQRCode==1.2.1
python-bitcoinrpc==1.0
python-dateutil==2.7.3
python-editor==1.0.3
python-oauth2==1.1.0
python3-keyczar==0.71rc0
python3-openid==3.1.0
pytz==2018.4
pyzmq==17.0.0
qtconsole==4.3.1
queuelib==1.5.0
rcssmin==1.0.6
redis==2.10.6
redo==1.6
requests==2.20.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rjsmin==1.0.12
rlp==1.0.1
s3transfer==0.1.13
scipy==1.0.0
Scrapy==1.5.0
scrapy-djangoitem==1.1.1
selenium==3.14.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.14.0
six==1.11.0
sorl-thumbnail==12.4.1
sortedcontainers==1.5.9
SQLAlchemy==1.2.7
sqlparse==0.2.4
statsmodels==0.8.0
TA-Lib==0.4.17
tables==3.4.2
tenacity==4.12.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
toolz==0.9.0
tornado==5.0.2
traitlets==4.3.2
Twisted==18.7.0
txaio==18.8.1
Unidecode==1.0.22
urllib3>=1.23
uvloop==0.11.0
uWSGI==2.0.17
vine==1.1.4
w3lib==1.19.0
wcwidth==0.1.7
web3==4.4.1
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.48.0
websockets==5.0.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
xlrd==1.1.0
yarl==1.1.0
zipline==1.2.0
zope.interface==4.5.0
